pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "emailfilter"
version = "0.1.0"
description = "Email filtering application"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
  "openai>=1.0.0",
  "python-dotenv>=1.0.0",
  "imapclient>=2.3.0",
  "email-validator>=2.0.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=23.0.0", "isort>=5.12.0", "mypy>=1.0.0"]

[project.scripts]
emailfilter = "emailfilter.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/emailfilter"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
